+++ pallets/swap-interface/src/order.rs
@@ -1,8 +1,15 @@
 use core::marker::PhantomData;
 
-use subtensor_runtime_common::Currency;
+use substrate_fixed::types::U64F64;
+use subtensor_runtime_common::{AlphaCurrency, Currency, TaoCurrency};
 
-pub struct OrderType<PaidIn, PaidOut>
+pub trait Order<PaidIn: Currency, PaidOut: Currency>: Clone {
+    fn amount(&self) -> PaidIn;
+    fn is_beyond_price_limit(&self, alpha_sqrt_price: U64F64, limit_sqrt_price: U64F64) -> bool;
+}
+
+#[derive(Clone)]
+pub struct BasicOrder<PaidIn, PaidOut>
 where
     PaidIn: Currency,
     PaidOut: Currency,
@@ -12,4 +19,22 @@ where
     _paid_out: PhantomData<PaidOut>,
 }
 
-pub trait Order<PaidIn: Currency, PaidOut: Currency> {}
+impl Order<TaoCurrency, AlphaCurrency> for BasicOrder<TaoCurrency, AlphaCurrency> {
+    fn amount(&self) -> TaoCurrency {
+        self.amount
+    }
+
+    fn is_beyond_price_limit(&self, alpha_sqrt_price: U64F64, limit_sqrt_price: U64F64) -> bool {
+        alpha_sqrt_price < limit_sqrt_price
+    }
+}
+
+impl Order<AlphaCurrency, TaoCurrency> for BasicOrder<AlphaCurrency, TaoCurrency> {
+    fn amount(&self) -> AlphaCurrency {
+        self.amount
+    }
+
+    fn is_beyond_price_limit(&self, alpha_sqrt_price: U64F64, limit_sqrt_price: U64F64) -> bool {
+        alpha_sqrt_price > limit_sqrt_price
+    }
+}
