+++ pallets/swap/src/pallet/mod.rs
@@ -17,6 +17,7 @@ use crate::{
 pub use pallet::*;
 
 mod impls;
+mod swap_step;
 #[cfg(test)]
 mod tests;
